requirements.txt
matplotlib==3.5.2
networkx==2.8.3
numpy==1.21.5
pandas==1.4.2
PyYAML==6.0
scipy==1.8.1
streamlit==1.13.0
torch==1.11.0+cu113
torch_geometric==2.1.0.post1
torch_sparse==0.6.15
